backend/requirements.txt
# --- Web framework ---
fastapi
uvicorn[standard]

# --- Database drivers & ORM ---
psycopg2-binary
asyncpg
sqlalchemy
pgvector

# --- Cache / Message broker ---
redis

# --- LangChain ecosystem ---
langchain
langchain-google-genai
langchain-core
langchain-community
langgraph

# --- Vector search ---
faiss-cpu

# --- ML / Embeddings ---
huggingface-hub
transformers
sentence-transformers

# --- Document parsing ---
pymupdf
python-docx

# --- Utilities ---
python-dotenv
pydantic
openai
